metric: fix windowing logic for manual histogram
These histograms were missing the windowing logic we use in histograms
to store them in our internal time series DB. The logic we follow there
is that we keep 3 histograms. The first one is a cumulative one that we
export to prometheus through the `status/vars` page. The second and
third are `prev` and `cur` that together cover the entire duration of
the histogram, default being 60s. When the ticker rotates these, we swap
out the prev with the current, and start a brand new current histogram.
When calculating derivitive statistics, we use the combined view of the
windows which is used to display these metrics in DB console. This is
similar to the way Grafana handles these using its rate interval.

Fixes #115825.

Release note: None